midterm/requirements.txt
Flask==2.3.3
flasgger==0.9.7.1
PyJWT==2.9.0
SQLAlchemy==1.4.54
